feat: unify preview component
